added SPI for STM32F0. changed delayus/delayms back to int for better performance on M0. added test project for STM32F0.

+++ HAL/STM32F0/F0SysTimer.h
@@ -10,7 +10,7 @@ namespace STM32F0
 		F0SysTimer();
 		~F0SysTimer(){}
 		virtual int64_t gettime();		// micro-second
-		virtual void delayms(float ms);
-		virtual void delayus(float us);
+		virtual void delayms(int ms);
+		virtual void delayus(int us);
 	};
 }
